make separation between internal and user settings clearer in code (fixes opacity regression)

+++ apps/VC/CVolumeViewerWithCurve.hpp
@@ -85,7 +85,7 @@ private:
     QCheckBox* fShowCurveBox;
     bool showCurve;
     int fwdBackMsJump;
-    float displaySegmentOpacity;
+    int displaySegmentOpacity;
     std::unordered_map<std::string, SegmentationStruct>& fSegStructMapRef;
     CBSpline* fSplineCurveRef;
     std::vector<cv::Vec2f> fControlPoints;
